fix: chart view - for client side filtering (#447)
