+++ src/components/Index/components/EntitiesView/components/EntityList/types.ts
@@ -1,3 +1,7 @@
-export interface EntityListProps {
+import { RowData, Table } from "@tanstack/react-table";
+
+export interface EntityListProps<T extends RowData> {
   entityListType: string;
+  loading: boolean;
+  table: Table<T>;
 }
