+++ src/components/Index/table/types.ts
@@ -1,0 +1,5 @@
+import { RowData, Table } from "@tanstack/react-table";
+
+export interface UseTable<T extends RowData> {
+  table: Table<T>;
+}
